3d-image-main/requirements.txt
torch
transformers
diffusers
pillow
matplotlib
numpy
pyrender
trimesh
opencv-python
rembg
